refactor: Компоонент MainScreen разбит на несколько файлов
