+++ src/components/main/main-offers.jsx
@@ -1,75 +1,13 @@
 import React from 'react';
-import Header from '../header/header';
-import Tabs from '../tabs/tabs';
-import PlacesSorting from '../places-sorting/places-sorting';
 import PropTypes from 'prop-types';
+import MainSorting from './main-sorting';
 
-const MainScreen = ({cities, offers, city, authorizedUser}) => {
-  const offersInCity = offers.filter((offer) => offer.city === city);
-  return (<div className="page page--gray page--main">
-    <Header isMain={true} authorizedUser={authorizedUser}/>
-    <main className={`page__main page__main--index${offersInCity.length > 0 ? `` : ` page__main--index-empty`}`}>
-      <h1 className="visually-hidden">Cities</h1>
-      <Tabs cities={cities} selectedCity={city}/>
-      <div className="cities">
-        {
-          offersInCity.length > 0
-            ? <OfferList offers={offersInCity} />
-            : <OfferEmpty city={city}/>
-        }
-      </div>
-    </main>
-  </div>);
-};
-
-MainScreen.propTypes = {
-  cities: PropTypes.arrayOf(
-      PropTypes.shape({
-        id: PropTypes.number.isRequired,
-        title: PropTypes.string.isRequired,
-        href: PropTypes.string.isRequired
-      })
-  ).isRequired,
-  offers: PropTypes.arrayOf(
-      PropTypes.shape({
-        id: PropTypes.number.isRequired,
-        title: PropTypes.string.isRequired,
-        city: PropTypes.string.isRequired,
-        type: PropTypes.string.isRequired,
-        price: PropTypes.number.isRequired,
-        rating: PropTypes.number.isRequired,
-        img: PropTypes.string.isRequired,
-        imgSmall: PropTypes.string.isRequired,
-        isPremium: PropTypes.bool,
-        isFavorite: PropTypes.bool
-      })
-  ).isRequired,
-  city: PropTypes.string.isRequired,
-  authorizedUser: PropTypes.string
-};
-
-const OfferEmpty = ({city}) => (
-  <div className="cities__places-container cities__places-container--empty container">
-    <section className="cities__no-places">
-      <div className="cities__status-wrapper tabs__content">
-        <b className="cities__status">No places to stay available</b>
-        <p className="cities__status-description">We could not find any property available at the moment in {city}</p>
-      </div>
-    </section>
-    <div className="cities__right-section"></div>
-  </div>
-);
-
-OfferEmpty.propTypes = {
-  city: PropTypes.string.isRequired
-};
-
-const OfferList = ({offers}) => (
+const MainOffers = ({offers}) => (
   <div className="cities__places-container container">
     <section className="cities__places places">
       <h2 className="visually-hidden">Places</h2>
       <b className="places__found">312 places to stay in Amsterdam</b>
-      <PlacesSorting />
+      <MainSorting />
       <div className="cities__places-list places__list tabs__content">
         {offers.map((offer) => <OfferCard key={`offerCard${offer.id}`} offer={offer} />)}
       </div>
@@ -80,7 +18,7 @@ const OfferList = ({offers}) => (
   </div>
 );
 
-OfferList.propTypes = {
+MainOffers.propTypes = {
   offers: PropTypes.arrayOf(
       PropTypes.shape({
         id: PropTypes.number.isRequired,
@@ -150,4 +88,4 @@ OfferCard.propTypes = {
   })
 };
 
-export default MainScreen;
+export default MainOffers;
